+++ src/PageInitialiser/init.ts
@@ -1,7 +1,3 @@
-import * as initLayout from './layout';
-import * as initDesign from './design';
-import * as initStations from './stations';
-import * as initInfo from './info';
 import { MDCTabBar } from '@material/tab-bar';
 import { MDCSlider } from '@material/slider';
 
@@ -10,40 +6,51 @@ export default function () {
 let firstInit = [false, true, true, true, true];
 window.sliders = [] as MDCSlider[];
 
-MDCTabBar.attachTo($('#panels .mdc-tab-bar')[0]).listen('MDCTabBar:activated', (event: any) => {
+MDCTabBar.attachTo($('#panels .mdc-tab-bar')[0]).listen('MDCTabBar:activated', (event: CustomEvent) => {
     $('.panel--active').removeClass('panel--active');
     $('.panel').eq(event.detail.index).addClass('panel--active');
 
     if (event.detail.index == 1 && firstInit[1]) {
-        initLayout.common();
-        if (window.urlParams.get('style') === 'gzmtr') {
-            initLayout.gzmtr();
-        }
-        firstInit[1] = false;
+        import(/* webpackChunkName: "initLayout" */ './layout')
+            .then(module => {
+                module.common();
+                if (window.urlParams.get('style') === 'gzmtr') {
+                    module.gzmtr();
+                }
+                firstInit[1] = false;
+            });
     }
     if (event.detail.index === 1) {
         window.sliders.forEach(slider => slider.layout());
     }
     if (event.detail.index == 2 && firstInit[2]) {
-        initDesign.common();
-        switch (window.urlParams.get('style')) {
-            case 'mtr':
-                initDesign.mtr();
-                break;
-            case 'gzmtr':
-                initDesign.gzmtr();
-                break;
-        }
-        firstInit[2] = false;
+        import(/* webpackChunkName: "initDesign" */ './design')
+            .then(module => {
+                module.common();
+                switch (window.urlParams.get('style')) {
+                    case 'mtr':
+                        module.mtr();
+                        break;
+                    case 'gzmtr':
+                        module.gzmtr();
+                        break;
+                }
+                firstInit[2] = false;
+            });
     }
     if (event.detail.index == 3 && firstInit[3]) {
-        console.log('init again');
-        initStations.common();
-        firstInit[3] = false;
+        import(/* webpackChunkName: "initStations" */ './stations')
+            .then(module => {
+                module.common();
+                firstInit[3] = false;
+            });
     }
     if (event.detail.index == 4 && firstInit[4]) {
-        initInfo.common();
-        firstInit[4] = false;
+        import(/* webpackChunkName: "initInfo" */ './info')
+            .then(module => {
+                module.common();
+                firstInit[4] = false;
+            });
     }
 });