+++ src/Station/StationGZ.ts
@@ -1,5 +1,46 @@
 import { RMGStation } from './Station';
-import { ID, Name, IntInfoTag, InterchangeInfo, BranchInfo, StationInfo, getIntBoxGZ, getTxtBoxDim } from '../utils';
+import { ID, Name, IntInfoTag, InterchangeInfo, BranchInfo, StationInfo } from '../utils';
+
+export function getIntBoxGZ(intInfo: InterchangeInfo, state: 0 | 1 | -1) {
+    let bg = intInfo[IntInfoTag.colour];
+    let fg = intInfo[IntInfoTag.fg];
+    let names = [
+        intInfo[IntInfoTag.nameZH], 
+        intInfo[IntInfoTag.nameEN]
+    ];
+    let nameZHs = names[0].match(/[\d]+|[\D]+/g) || [''];
+    let intNameSplitOk = false;
+    if (nameZHs.length == 2) {
+        if (!isNaN(Number(nameZHs[0])) && isNaN(Number(nameZHs[1]))) {
+            intNameSplitOk = true;
+        }
+    }
+    let boxEl = $('<g>')
+        .append(
+            $('<use>', { 
+                'xlink:href': '#intbox_gz', 
+                fill: state===-1 ? '#aaa' : bg
+            })
+        )
+        .append(
+            $('<text>', { y: 8.5, class: 'rmg-name__zh rmg-name__gzmtr--int' })
+                .append($('<tspan>', { 'font-size':'16px', 'dominant-baseline': 'central' }).text(intNameSplitOk ? nameZHs[0] : ''))
+                .append($('<tspan>', { dy:-1, 'dominant-baseline': 'central' }).text(intNameSplitOk ? nameZHs[1] : nameZHs.join('')))
+        )
+        .append(
+            $('<text>', { 
+                y: 19.5, 
+                class: 'rmg-name__en'
+            })
+                .addClass(names[1].length > 10 ? 'rmg-name__gzmtr--int-small' : 'rmg-name__gzmtr--int')
+                .text(names[1])
+        );
+    if (fg === '#fff' || state === -1) {
+        $(boxEl).find('text').addClass('rmg-name__gzmtr--white-fg');
+    }
+
+    return boxEl;
+}
 
 class RMGStationGZ extends RMGStation {
     constructor(id: ID, data: StationInfo) {
@@ -67,7 +108,7 @@ class RMGStationGZ extends RMGStation {
         }
         let stnNameDim = ($(`#stn_icons #${this.id} g#stn_name g`)[0] as Element as SVGGElement).getBBox();
         return $('<g>', { 
-            transform: `translate(${(stnNameDim.width+35)*(this._tickRotation===0?1:-1)},${2.5+5*(this.name[1].split('\\').length-1)})`, 
+            transform: `translate(${(stnNameDim.width+35)*(this._tickRotation===0?1:-1)},${2+5*(this.name[1].split('\\').length-1)})`, 
             fill: this.state===-1 ? '#aaa' : colour, 
             'text-anchor': 'middle'
         })